+++ packages/showcases/css/stories/components/selection-controls/quantity/quantity.stories.js
@@ -1,8 +1,11 @@
 import overview from './examples/overview.html';
 
+import { parameters } from '@vtmn/showcase-core/csf/components/selection-controls/quantity.csf';
+
 export default {
   title: 'Components / Selection controls / Quantity',
   argTypes: { onclick: { action: 'clicked' } },
+  parameters,
 };
 
 export const Overview = () => overview;