+++ packages/showcases/react/stories/components/actions/VtmnLink/VtmnLink.stories.tsx
@@ -1,0 +1,28 @@
+import React from 'react';
+import { VtmnLink } from '@vtmn/react';
+import { Meta, Story } from '@storybook/react';
+import {
+  argTypes,
+  parameters,
+} from '@vtmn/showcase-core/csf/components/actions/link.csf';
+
+export default {
+  title: 'Components / Actions / VtmnLink',
+  component: VtmnLink,
+  argTypes: {
+    ...argTypes,
+    children: {
+      type: { name: 'string', required: false },
+      description: 'The content to render inside the component.',
+      defaultValue: 'Link',
+      control: {
+        type: 'text',
+      },
+    },
+  },
+  parameters,
+} as Meta;
+
+const Template: Story = (args) => <VtmnLink {...args} />;
+
+export const Overview = Template.bind({});