+++ packages/showcases/vue/.storybook/preview.js
@@ -49,13 +49,9 @@ export const parameters = {
   viewport: {
     viewports,
   },
-  docs: {
-    transformSource(src, ctx) {
-      const match = /\b("')?template\1:\s*`([^`]+)`/.exec(src);
-      if (match) {
-        return templateSourceCode(dedent(match[2]), ctx.args, ctx.argTypes);
-      }
-      return src;
+  options: {
+    storySort: {
+      order: ['Guidelines', 'Components'],
     },
-  }
+  },
 };