+++ packages/sources/css/src/components/indicators/badge/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.10.4](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-badge@0.10.3...@vtmn/css-badge@0.10.4) (2022-02-08)
+
+**Note:** Version bump only for package @vtmn/css-badge
+
+
+
+
+
+## [0.10.3](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-badge@0.10.2...@vtmn/css-badge@0.10.3) (2022-02-03)
+
+**Note:** Version bump only for package @vtmn/css-badge
+
+
+
+
+
+## [0.10.2](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-badge@0.10.1...@vtmn/css-badge@0.10.2) (2022-02-02)
+
+**Note:** Version bump only for package @vtmn/css-badge
+
+
+
+
+
+## [0.10.1](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-badge@0.10.0...@vtmn/css-badge@0.10.1) (2022-01-28)
+
+**Note:** Version bump only for package @vtmn/css-badge
+
+
+
+
+
 # [0.10.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-badge@0.9.0...@vtmn/css-badge@0.10.0) (2022-01-20)