+++ packages/sources/css/src/components/overlays/tooltip/CHANGELOG.md
@@ -3,6 +3,49 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.4](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-tooltip@0.6.3...@vtmn/css-tooltip@0.6.4) (2022-02-08)
+
+**Note:** Version bump only for package @vtmn/css-tooltip
+
+
+
+
+
+## [0.6.3](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-tooltip@0.6.2...@vtmn/css-tooltip@0.6.3) (2022-02-03)
+
+**Note:** Version bump only for package @vtmn/css-tooltip
+
+
+
+
+
+## [0.6.2](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-tooltip@0.6.1...@vtmn/css-tooltip@0.6.2) (2022-02-02)
+
+**Note:** Version bump only for package @vtmn/css-tooltip
+
+
+
+
+
+## [0.6.1](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-tooltip@0.6.0...@vtmn/css-tooltip@0.6.1) (2022-01-28)
+
+**Note:** Version bump only for package @vtmn/css-tooltip
+
+
+
+
+
+# [0.6.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-tooltip@0.5.0...@vtmn/css-tooltip@0.6.0) (2022-01-24)
+
+
+### Features
+
+* **@vtmn/css:** remove outline when components are focus on mobile devices ([#931](https://github.com/Decathlon/vitamin-web/issues/931)) ([719a379](https://github.com/Decathlon/vitamin-web/commit/719a379f5863d9c95e67795fb1ca0e4975e12a4e))
+
+
+
+
+
 # [0.5.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-tooltip@0.4.10...@vtmn/css-tooltip@0.5.0) (2022-01-20)