+++ packages/sources/css/src/components/structure/accordion/CHANGELOG.md
@@ -3,6 +3,49 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.4](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-accordion@0.2.3...@vtmn/css-accordion@0.2.4) (2022-02-08)
+
+**Note:** Version bump only for package @vtmn/css-accordion
+
+
+
+
+
+## [0.2.3](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-accordion@0.2.2...@vtmn/css-accordion@0.2.3) (2022-02-03)
+
+**Note:** Version bump only for package @vtmn/css-accordion
+
+
+
+
+
+## [0.2.2](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-accordion@0.2.1...@vtmn/css-accordion@0.2.2) (2022-02-02)
+
+**Note:** Version bump only for package @vtmn/css-accordion
+
+
+
+
+
+## [0.2.1](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-accordion@0.2.0...@vtmn/css-accordion@0.2.1) (2022-01-28)
+
+**Note:** Version bump only for package @vtmn/css-accordion
+
+
+
+
+
+# [0.2.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-accordion@0.1.1...@vtmn/css-accordion@0.2.0) (2022-01-21)
+
+
+### Features
+
+* **@vtmn/css:** remove acccordion list-type on safari browser and other non supported ([#920](https://github.com/Decathlon/vitamin-web/issues/920)) ([fff8e29](https://github.com/Decathlon/vitamin-web/commit/fff8e2913766fbd8eb82d8ad1d4b33455195119c))
+
+
+
+
+
 ## [0.1.1](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-accordion@0.1.0...@vtmn/css-accordion@0.1.1) (2022-01-20)